[Feat] 다른 사람의 프로필 화면 UI 구현

+++ core/network/src/main/java/com/ilsangtech/ilsang/core/network/model/user/TopCommercialAreaNetworkModel.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class TopCommercialAreaNetworkModel(
     val commercialAreaCode: String,
-    val ownerContributionPercent: Int,
+    val ownerContributionPercent: Double,
     val point: Int
 )
